pyproject.toml
[project]
name = "akeneo-client"
version = "0.1.0"
description = "Simple Akeneo PIM API client (Connector flow)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "pydantic>=2.8",
  "typer[all]>=0.12",
  "python-dotenv>=1.0",
  "tenacity>=9.0",
  "rich>=13.7"
]

[project.scripts]
akeneo = "akeneo_client.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
